Move vector utils into common
